pyproject.toml
[tool.poetry]
name = "el-xizmati"
version = "0.1.0"
description = ""
authors = ["zukhriddin <zux.828@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
dj-database-url = "^2.2.0"
django-redis = "^5.4.0"
django-celery-beat = "^2.6.0"
django-celery-results = "^2.5.1"
gunicorn = "^22.0.0"
fcm-django = "^2.1.0"
django-extra-settings = "^0.12.0"
django-filter = "^24.2"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
environs = "^11.0.0"
djangorestframework = "^3.15.2"
flower = "^2.0.1"
pillow = "^10.4.0"
drf-spectacular = "^0.27.2"
drf-api-tracking = "^1.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
